Revert "Lots of tests"
