+++ .taprc
@@ -2,7 +2,6 @@ branches: 100
 check-coverage: true
 color: true
 coverage: true
-reporter: spec
 coverage-report:
   - text
   - cobertura
