+++ package.json
@@ -12,6 +12,7 @@
     "start": "docker-compose up -d --build",
     "stop": "docker-compose down",
     "test": "npm run lint-test",
+    "tap": "tap --coverage test",
     "doc": "jsdoc -R README.md -P package.json -r src",
     "migrate": "pg-migrator postgres://postgres:mco@localhost/postgres 2",
     "prod": "node src/app.js | ./node_modules/.bin/bunyan",
@@ -38,8 +39,6 @@
     "c-struct": "0.0.5",
     "pg": "8.5.1",
     "pg-migrator": "1.0.5",
-    "snyk": "1.431.4",
-    "statsd-client": "0.4.5"
     "snyk": "1.434.4",
     "statsd-client": "0.4.5",
     "ts-node": "9.1.1",
@@ -50,7 +49,6 @@
     "@types/node": "^14.14.10",
     "jsdoc": "3.6.6",
     "standard": "^16.0.3",
-    "supertest": "6.0.1",
     "tap": "14.11.0"
   },
   "jest": {
