+++ src/app.js
@@ -7,12 +7,11 @@
 
 const { PatchServer } = require('./services/PatchAndShard/patchServer')
 const { Server } = require('./server')
-const { Logger } = require('./services/shared/loggerManager')
 
 // MCOS Monolith
 const server = new Server('./src/services/shared/config.json')
 server.start()
 
 // MCOS PatchAndShard
-const patchAndShardServer = new PatchServer(new Logger().getLogger('PatchServer'))
+const patchAndShardServer = new PatchServer('./src/services/shared/config.json')
 patchAndShardServer.start()
