+++ src/services/shared/databaseManager.js
@@ -5,6 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
+const { Logger } = require('./loggerManager')
 const pool = require('./db/index')
 
 /**
@@ -13,10 +14,9 @@ const pool = require('./db/index')
 class DatabaseManager {
   /**
    *
-   * @param {Logger} logger
    */
-  constructor (logger) {
-    this.logger = logger
+  constructor () {
+    this.logger = new Logger().getLogger('DatabaseManager')
   }
 
   /**
@@ -34,7 +34,6 @@ class DatabaseManager {
       return rows[0]
     } catch (e) {
       this.logger.warn(`Unable to update session key ${e}`)
-      throw new Error(e)
     }
   }
